feat: extract common base class for ask ai widget
